+++ src/app/admin/layout.tsx
@@ -1,24 +1,16 @@
-import React from "react";
-import Sidebar from "./components/Sidebar";
-import { CoursesProvider } from "./context/CoursesContext";
 import { Toaster } from "sonner";
 
-export default function AdminLayout({
+export default function DashboardLayout({
   children,
 }: {
-  children: React.ReactNode;
+  children: React.ReactNode
 }) {
   return (
-    <CoursesProvider>
-      <div className="flex">
-        <div className='fixed'>
-        <Sidebar />
-        </div>
-        <main className="flex-1 bg-gray-50 min-h-screen ml-[240px]">
-          {children}
-          <Toaster position="bottom-right" richColors />
-        </main>
-      </div>
-    </CoursesProvider>
-  );
+    <html lang="en">
+      <body>
+        <main>{children}</main>
+        <Toaster position="bottom-right" richColors />
+      </body>
+    </html>
+  )
 }
